+++ loan-broker-commons-db/src/main/java/com/loanbroker/commons/db/ResultRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResultRepository extends CrudRepository<Result, Integer> {
+public interface ResultRepository extends CrudRepository<Result, Long> {
 }
